pyproject.toml
[project]
name = "scix-experiments"
version = "0.1.0"
description = "AI/ML experiments on scientific literature — making knowledge navigable by agents"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "psycopg[binary]>=3.1",
    "psycopg-pool>=3.1",
    "wordfreq>=3.1",
    "pyahocorasick>=2.0",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "black>=24.0",
    "ruff>=0.3",
    "libcst>=1.1",
    "hypothesis>=6.100",
]
embed = [
    "transformers>=4.36",
    "torch>=2.1",
]
search = [
    "sentence-transformers>=2.6",
]
mcp = [
    "mcp>=1.2,<2.0",
]
analysis = [
    "matplotlib>=3.7",
]
graph = [
    "python-igraph>=0.11",
    "leidenalg>=0.10",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "tests"]
markers = [
    "integration: tests requiring a running scix database",
]
